[LEARNING]-SpringBoot: CRUD application with Spring Data REST (End of Section 4).
1) Spring Data REST dependency will scan for JPARepositories interfaces and expose the default CRUD endpoints like a real REST controller (wow);

2) Spring Data REST allows configurations like base-path, default page size (pagination), max page size and more;

3) When you extend JpaRepository, Spring Data JPA automatically creates a proxy implementation of the repository interface at runtime. Spring will automatically scan for interfaces that extend JpaRepository (or CrudRepository, PagingAndSortingRepository) and create the necessary beans for them, without requiring the @Repository annotation.

However:
Exception Translation: One of the main reasons to use @Repository is for Spring's automatic translation of database exceptions into Spring's DataAccessException. If you don't annotate your repository interface or class with @Repository, you might not get this exception translation in some cases. But even this is not strictly required in most cases, as Spring will apply it if your repository is managed properly by the framework.
In summary, while the @Repository annotation is not necessary for Spring Data JPA repositories to work, it is still considered good practice because it clarifies the role of the class or interface and ensures consistent exception translation across layers;

4) @RepositoryRestResource - used to change the exposed resource endpoint's name from employees (default behavior from Spring Data REST, entity is "Employee" which translates the resource name to "employees") to members (custom name set);

5) Default pagination is set to 3 via application.properties configuration but it can be configured through the URL as shown below:
http://localhost:8080/api/members?page=1&size=2

6) Sort can also be specified as shown below:
http://localhost:8080/api/members?sort=lastName,desc